feat: Introduce Compactor service for context compaction and XML tool call handling
- Added `Compactor` service to handle context compaction based on agent settings.
- Implemented methods for compressing sequences of messages and generating summaries.
- Created utility functions for extracting and removing XML tags.
- Updated `Orchestrator` to integrate compaction logic during chat turns.
- Introduced error handling for tool call arguments using `thiserror`.
- Added new dependencies: `async-recursion`, `chrono`, `derive_setters`, `forge_walker`, `handlebars`, `pretty_assertions`, `regex`, `rust-embed`, `serde_json`, `tokio`, and `thiserror`.
- Refactored existing code to accommodate new features and improve modularity.
