+++ crates/forge_app/Cargo.toml
@@ -8,3 +8,18 @@ forge_domain.workspace = true
 async-trait.workspace = true
 anyhow.workspace = true
 serde.workspace = true
+futures.workspace = true
+tracing.workspace = true
+async-recursion.workspace = true
+chrono.workspace = true
+derive_setters.workspace = true
+forge_walker.workspace = true
+handlebars.workspace = true
+rust-embed.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+regex.workspace = true
+thiserror.workspace = true
+
+[dev-dependencies]
+pretty_assertions.workspace = true
