+++ crates/forge_app/src/error.rs
@@ -1,0 +1,5 @@
+#[derive(thiserror::Error, Debug)]
+pub enum Error {
+    #[error("Invalid tool call arguments: {0}")]
+    ToolCallArgument(serde_json::Error),
+}
