+++ crates/forge_app/src/lib.rs
@@ -1,3 +1,10 @@
+mod compaction;
+mod error;
+mod orch;
 mod services;
+mod utils;
 
+pub use compaction::*;
+pub use error::*;
+pub use orch::*;
 pub use services::*;
