+++ crates/forge_domain/src/lib.rs
@@ -4,7 +4,6 @@ mod attachment;
 mod chat_request;
 mod chat_response;
 mod compact;
-mod compaction;
 mod compaction_result;
 mod conversation_html;
 mod update;
@@ -20,7 +19,6 @@ mod mcp;
 mod merge;
 mod message;
 mod model;
-mod orch;
 mod point;
 mod provider;
 mod retry_config;
@@ -29,7 +27,6 @@ mod suggestion;
 mod system_context;
 mod temperature;
 mod template;
-mod text_utils;
 mod tool;
 mod tool_call;
 mod tool_call_context;
@@ -50,7 +47,6 @@ pub use attachment::*;
 pub use chat_request::*;
 pub use chat_response::*;
 pub use compact::*;
-pub use compaction::*;
 pub use compaction_result::*;
 pub use context::*;
 pub use conversation::*;
@@ -63,7 +59,6 @@ pub use image::*;
 pub use mcp::*;
 pub use message::*;
 pub use model::*;
-pub use orch::*;
 pub use point::*;
 pub use provider::*;
 pub use retry_config::*;
@@ -72,7 +67,6 @@ pub use suggestion::*;
 pub use system_context::*;
 pub use temperature::*;
 pub use template::*;
-pub use text_utils::*;
 pub use tool::*;
 pub use tool_call::*;
 pub use tool_call_context::*;
